platform-checks: Add timestamp oracle toggle scenario
Part of https://www.notion.so/materialize/platform-v2-Distributed-Timestamp-Oracle-fb625f177eeb40858dc4324e78914da6
